+++ .gitignore
@@ -6,6 +6,7 @@ __pycache__/
 # C extensions
 *.so
 *.a
+*.dll
 wolfcrypt/_ffi.*
 
 # Distribution / packaging
