+++ MANIFEST.in
@@ -2,6 +2,10 @@ include LICENSING.rst
 include README.rst
 include ChangeLog.rst
 
+include scripts/build_ffi.py
+include windows/fips_ready/user_settings.h
+include windows/non_fips/user_settings.h
+
 recursive-include tests *
 recursive-include requirements *
